requirements.txt
requests==2.31.0
environs==9.5.0
web3==6.8.0
python-dotenv~=1.0.0
loguru~=0.7.0
fake-useragent==1.2.1
openpyxl~=3.1.2